+++ components/atoms/Button.js
@@ -1,4 +1,3 @@
-import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
@@ -54,41 +53,48 @@ const StyledButton = styled.button`
 
   &:hover {
     border: ${props => {
-      if (props.hollow) {
-        return `${px(props.border)} solid ${props.inverted ? props.theme.colors.gray1 : props.theme.colors.blue4}`;
-      }
-      return 'none'
-    }};
+    if (props.hollow) {
+      return `${px(props.border)} solid ${props.inverted ? props.theme.colors.gray1 : props.theme.colors.blue4}`
+    }
+    return 'none'
+  }};
     background-color: ${props => {
-      if (!props.hollow) {
-        return props.inverted ? props.theme.colors.gray1 : props.theme.colors.blue4
-      }
-      return 'transparent'
-    }};
+    if (!props.hollow) {
+      return props.inverted ? props.theme.colors.gray1 : props.theme.colors.blue4
+    }
+    return 'transparent'
+  }};
   }
   &:active {
     transition: .2s ease-in;
     border: ${props => {
-      if (props.hollow) {
-        return `${px(props.border)} solid ${props.inverted ? props.theme.colors.gray2 : props.theme.colors.blue6}`;
-      }
-      return 'none'
-    }};
+    if (props.hollow) {
+      return `${px(props.border)} solid ${props.inverted ? props.theme.colors.gray2 : props.theme.colors.blue6}`
+    }
+    return 'none'
+  }};
     background-color: ${props => {
-      if (!props.hollow) {
-        return props.inverted ? props.theme.colors.gray2 : props.theme.colors.blue6;
-      }
-      return 'transparent'
-    }};
+    if (!props.hollow) {
+      return props.inverted ? props.theme.colors.gray2 : props.theme.colors.blue6
+    }
+    return 'transparent'
+  }};
   }
 `
 
-export const Button = StyledButton.extend`
+export const Button = styled(StyledButton)`
   ${space}
   ${width}
   ${fontSize}
   ${color}
 `
+Button.propTypes = {
+  fontSize: PropTypes.number,
+  inverted: PropTypes.bool,
+  hollow: PropTypes.bool,
+  border: PropTypes.number
+}
+
 Button.defaultProps = {
   fontSize: 1,
   inverted: false,