+++ components/atoms/Radio.js
@@ -1,11 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import { Flex, Box } from 'grid-styled'
+import { Flex, Box } from '@rebass/grid'
 
-import MdWarning from 'react-icons/lib/md/warning'
-
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 
 import { space, fontSize, fontSizeMult, color, width } from '../util'
 
@@ -80,14 +78,14 @@ const StyledRadioContainer = styled.div`
   ${fontSizeMult('width', 1)}
 `
 
-const StyledRadio = StyledRadioContainer.extend`
+const StyledRadio = styled(StyledRadioContainer)`
   border: 2px solid ${ props => props.checked ? props.theme.colors.base : props.theme.colors.gray5 };
   background-color: 'transparent';
   position: absolute;
   top: 0;
 `
 
-const StyledRadioInner = StyledRadio.extend`
+const StyledRadioInner = styled(StyledRadio)`
   background-color: ${ props => props.theme.colors.base};
   transform: scale(0.5);
 `
@@ -107,24 +105,23 @@ export class RadioButton extends React.Component {
 
   blur() {
     if (this.inputNode) {
-      this.inputNode.blur();
+      this.inputNode.blur()
     }
   }
 
   focus() {
     if (this.inputNode) {
-      this.inputNode.focus();
+      this.inputNode.focus()
     }
   }
 
-
-  renderLabel(props) {
-    if (React.isValidElement(props.label)) {
-      return React.cloneElement(props.label, props)
+  renderLabel(label, props) {
+    if (React.isValidElement(label)) {
+      return React.cloneElement(label, props)
     }
     return (
       <div>
-        {props.label}
+        {label}
       </div>
     )
   }
@@ -135,20 +132,20 @@ export class RadioButton extends React.Component {
       <StyledRadioButton {...this.props}>
         <StyledLabel {...this.props}>
           <Flex>
-          <Box>
-            <ErasedInputRadio
-              onClick={this.handleClick}
-              onChange={()  => {}}
-              innerRef={c => {this.inputNode = c}}
-              type='radio' />
-            <StyledRadioContainer {...this.props}>
-              <StyledRadio checked={checked} {...this.props}/>
-              {checked && <StyledRadioInner {...this.props} />}
-            </StyledRadioContainer>
-          </Box>
-          <Box>
-            {this.renderLabel(this.props)}
-          </Box>
+            <Box>
+              <ErasedInputRadio
+                onClick={this.handleClick}
+                onChange={()  => {}}
+                innerRef={c => {this.inputNode = c}}
+                type='radio' />
+              <StyledRadioContainer {...this.props}>
+                <StyledRadio checked={checked} {...this.props}/>
+                {checked && <StyledRadioInner {...this.props} />}
+              </StyledRadioContainer>
+            </Box>
+            <Box>
+              {this.renderLabel(label, this.props)}
+            </Box>
           </Flex>
         </StyledLabel>
       </StyledRadioButton>